requirements.txt
lxml == 4.* 
PyMySQL==0.10.1
pytz==2020.4
requests==2.25.0
PyJWT==1.7.1
python-dotenv==0.15